Add jebbs.plantuml
By trial-and-error I came to the conclusion that this extension requires one to
add an official "plantuml.jar" in the extension's root, that will be used for
local, serverless rendering. This apparently does not happen automatically.

That plantuml.jar is available under a few different licenses. I chosed the MIT
one, to match this extension's license. When the zip file is extracted a "COPYING"
file is created along with plantuml.jar, that documents its provenance and license.
For clarity I rename this file to COPYING.plantuml-jar and leave it in place to be
packaged along with the extension.

Here's where the URL used to fetch plantuml.jar comes from:

Start at the official PlantUML download page:
https://plantuml.com/download
Chose the MIT-version at the bottom of the page and copy the sourceforge link.
At the time of writing it is:
http://sourceforge.net/projects/plantuml/files/plantuml-jar-mit-1.2021.0.zip/download

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>