Add Jupyter extension
This extension is a dependency for recent versions of ms-python.python.
Adding this should fix the publishing error for the Python extension.

The prepublish step is taken from the extension's CONTRIBUTE.md file.
PIP_USER=no is added to prevent its presence otherwise to intefere with
-t flag.

ms-toolsai.jupyter is available under MIT license.