Fail for extensions above 10M
This changes logic used for determining whether we fail the workflow.
Previously, it was a lot of magic looking at the results from the last / day week,
but now it's a hard bar of reporting only if an extension has 10+ milion downloads
on the Microsoft Marketplace.
In addition, this formats the `report-extensions` file